+++ src/switcher/main-window/TileView.swift
@@ -311,7 +311,7 @@ class TileView: FlippedView {
 
     private func updateAppIcon(_ element: Window, _ title: String) {
         let appIconSize = TileView.iconSize()
-        appIcon.updateContents(.cgImage(element.icon), appIconSize)
+        appIcon.updateContents(.cgImage(SafariSiteIcons.icon(for: element) ?? element.icon), appIconSize)
     }
 
     private func updateValues(_ element: Window, _ index: Int, _ newHeight: CGFloat) {